extract guess_value from arxml to utils module create test for guess_value fix guess_value for python2
